requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
ffmpy==0.2.2
idna==2.6
PyYAML==3.12
qcloud-cos-v4==0.0.22
requests==2.18.4
urllib3==1.22
